requirements.txt
langchain
langchain-community
langchain_openai
duckduckgo-search
python-dotenv
openai
langgraph
streamlit